Add AbstractConfiguration and ConfigurationProvider
Move the shared implementation of configuration handling into a new AbstractConfiguration class.

Add a Container ConfigurationProvider and unit tests
